Use docker-compose for environment setup
Using docker-compose allows us to configure more environment aspects
(e.g. bind mounts or environment variables) with the docker-compose.yml
file.

The workspace path is different than in the Dockerfile-based setup, but
this is actually better: now we have complete control over the workspace
path, so we set it to be the same as in CI usage.

CMD is required during Dev Containers startup: without a long-running
command the container exits before Dev Containers extension can attach
to it during container setup phase. "sleep infinity" appears to be a
commonly used workaround.